+++ CHANGELOG.rst
@@ -10,7 +10,8 @@ Change Log
 =====
 * 2023-07-17
 * Support local-check-runner utility.
-* Minor change to respect REDIS_HOST_LOCAL environment variable (for local dev/testing).
+* Minor change to respect REDIS_HOST_LOCAL environment variable (for local dev/testing),
+  as well as allowing override of Auth0 client/secret (AUTH0_CLIENT_LOCAL, AUTH0_SECRET_LOCAL).
 
 
 4.4.0